Simplify some more (too much?)
